+++ sparcal_viewer/__init__.py
@@ -2,7 +2,7 @@
 import datetime
 import os
 
-__version__ = "1.4.0"
+__version__ = "1.5.0"
 
 
 def build_time() -> str: